refactor(Chat + Papa): too much changed - next time smaller commits :)
